feat(HList): Totally move HeterogeneousList in new file datatools, which also contains CategorizedList. File cluster.py deleted
